requirements.txt
flask
crewai
agentops
langsmith
langchain-community
langchain-openai
pywin32